Increase sizeof frameAirTimeUsec to handle > 65ms frames.

+++ Firmware/LoRaSerial_Firmware/LoRaSerial_Firmware.ino
@@ -376,7 +376,7 @@ bool cylonPatternGoingLeft;
 //-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 uint16_t radioBand = 0; //In MHz. Detected radio module type. Sets the upper/lower frequency bounds.
 uint8_t outgoingPacket[MAX_PACKET_SIZE]; //Contains the current data in route to receiver
-uint16_t frameAirTimeUsec; //Recalc'd with each new packet transmission
+uint32_t frameAirTimeUsec; //Recalc'd with each new packet transmission
 uint8_t frameSentCount = 0; //Increases each time a frame is sent
 
 unsigned long lastPacketReceived = 0; //Controls link LED in broadcast mode
